HN_case_study/requirements.txt
APScheduler
asgiref
backports.zoneinfo
Django
django-apscheduler
django-cors-headers
djangorestframework
djangorestframework-simplejwt
dnspython
mongoengine
mysql-connector
mysql-connector-python
mysqlclient
PyJWT
pymongo
pytz
six
sqlparse
typing-extensions
tzlocal
